+++ packages/components/combobox/src/combobox.ts
@@ -1,6 +1,6 @@
 import { LOCALE_STATUS_EVENT, localized, msg, str } from '@lit/localize';
 import { type ScopedElementsMap, ScopedElementsMixin } from '@open-wc/scoped-elements/lit-element.js';
-import { FormControlMixin, type SlUpdateStateEvent } from '@sl-design-system/form';
+import { FormControlMixin, type SlFormControlEvent, type SlUpdateStateEvent } from '@sl-design-system/form';
 import { Icon } from '@sl-design-system/icon';
 import { Option, OptionGroup } from '@sl-design-system/listbox';
 import { type EventEmitter, EventsController, anchor, event } from '@sl-design-system/shared';
@@ -14,6 +14,7 @@ import { repeat } from 'lit/directives/repeat.js';
 import styles from './combobox.scss.js';
 import { CreateCustomOption } from './create-custom-option.js';
 import { CustomOption } from './custom-option.js';
+import { NoMatch } from './no-match.js';
 import { SelectedGroup } from './selected-group.js';
 
 declare global {
@@ -57,6 +58,7 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
     return {
       'sl-combobox-create-custom-option': CreateCustomOption,
       'sl-combobox-custom-option': CustomOption,
+      'sl-combobox-no-match': NoMatch,
       'sl-combobox-selected-group': SelectedGroup,
       'sl-icon': Icon,
       'sl-option': Option,
@@ -73,7 +75,10 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
   static viewportMargin = 8;
 
   /** Event controller. */
-  #events = new EventsController(this);
+  #events = new EventsController(this, { click: this.#onClick });
+
+  /** Message element for when filtering results did not yield any results. */
+  #noMatch?: NoMatch;
 
   /** Monitor the DOM for new options. */
   #observer = new MutationObserver(() => this.#updateOptions());
@@ -278,6 +283,7 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
         @sl-blur=${this.#onTextFieldBlur}
         @sl-change=${this.#onTextFieldChange}
         @sl-focus=${this.#onTextFieldFocus}
+        @sl-form-control=${this.#onTextFieldFormControl}
         @sl-update-state=${this.#onTextFieldUpdateState}
         ?disabled=${this.disabled}
         ?readonly=${this.selectOnly}
@@ -356,6 +362,12 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
   #onButtonClick(): void {
     if (!this.#popoverJustClosed) {
       this.wrapper?.showPopover();
+    }
+  }
+
+  #onClick(event: Event): void {
+    if (event.target === this) {
+      this.input.focus();
     }
   }
 
@@ -533,6 +545,11 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
     this.focusEvent.emit();
   }
 
+  #onTextFieldFormControl(event: SlFormControlEvent): void {
+    event.preventDefault();
+    event.stopPropagation();
+  }
+
   #onTextFieldUpdateState(event: SlUpdateStateEvent): void {
     event.preventDefault();
     event.stopPropagation();
@@ -658,16 +675,31 @@ export class Combobox<T = unknown> extends FormControlMixin(ScopedElementsMixin(
   }
 
   #updateFilteredOptions(value?: string): void {
+    let noMatch = true;
+
     this.options.forEach(option => {
       let match = !this.filterResults || !value;
       if (!match) {
         match = option.content.toLowerCase().startsWith(value!.toLowerCase());
       }
 
+      if (noMatch && match) {
+        noMatch = false;
+      }
+
       if (option.element) {
         option.element.style.display = match ? '' : 'none';
       }
     });
+
+    if (noMatch && value) {
+      this.#noMatch ||= this.shadowRoot!.createElement('sl-combobox-no-match') as NoMatch;
+      this.#noMatch.value = value;
+      this.listbox?.prepend(this.#noMatch);
+    } else {
+      this.#noMatch?.remove();
+      this.#noMatch = undefined;
+    }
   }
 
   /** Updates the list of options and the listbox link with the text input. */